feat(workflow): Adding functionality to enable+disable alert rules (#19717)
* Adding functionality to enable + disable alert rules

* Adding disabled function

* Adding test around disable, fixing docstring

* feat(workflow): Adding functionality to enable alert rules (#19762)

* Tidying

* Fixing bug where subscription id was nulled too early

* Fixing nit

+++ src/sentry/snuba/tasks.py
@@ -49,7 +49,6 @@ def create_subscription_in_snuba(query_subscription_id, **kwargs):
     if subscription.subscription_id is not None:
         metrics.incr("snuba.subscriptions.create.already_created_in_snuba")
         return
-
     subscription_id = _create_in_snuba(subscription)
     subscription.update(
         status=QuerySubscription.Status.ACTIVE.value, subscription_id=subscription_id
@@ -97,7 +96,9 @@ def update_subscription_in_snuba(query_subscription_id, old_dataset=None, **kwar
 def delete_subscription_from_snuba(query_subscription_id, **kwargs):
     """
     Task to delete a corresponding subscription in Snuba from a `QuerySubscription` in
-    Sentry. Deletes the local subscription once we've successfully removed from Snuba.
+    Sentry.
+    If the local subscription is marked for deletion (as opposed to disabled),
+    then we delete the local subscription once we've successfully removed from Snuba.
     """
     try:
         subscription = QuerySubscription.objects.get(id=query_subscription_id)
@@ -105,7 +106,10 @@ def delete_subscription_from_snuba(query_subscription_id, **kwargs):
         metrics.incr("snuba.subscriptions.delete.subscription_does_not_exist")
         return
 
-    if subscription.status != QuerySubscription.Status.DELETING.value:
+    if subscription.status not in [
+        QuerySubscription.Status.DELETING.value,
+        QuerySubscription.Status.DISABLED.value,
+    ]:
         metrics.incr("snuba.subscriptions.delete.incorrect_status")
         return
 
@@ -114,7 +118,10 @@ def delete_subscription_from_snuba(query_subscription_id, **kwargs):
             QueryDatasets(subscription.snuba_query.dataset), subscription.subscription_id
         )
 
-    subscription.delete()
+    if subscription.status == QuerySubscription.Status.DELETING.value:
+        subscription.delete()
+    else:
+        subscription.update(subscription_id=None)
 
 
 def build_snuba_filter(dataset, query, aggregate, environment, params=None):
